services/requirements.txt
omdbpy
requests
